Some experiments with the job system. Added additional vertex streams. Loaded a whole Sponza scene. Cleanup Application.cpp, moved all scene loading into a separate class. Minor fixes. Broke scene saves and loads.

+++ Engine/RenderInterface/Source/RenderInterfaceCommonTypes.h
@@ -221,7 +221,7 @@ namespace toy::renderer
 	{
 		Handle<Buffer> buffer;
 		core::u32 offset{};
-		core::u32 size{};
+		core::u64 size{};
 	};
 
 	struct CBV
